refactor(web): extract VersionDiffView from VersionDiffModal so the renderer can be reused
The diff-rendering layout (Modified / Added / Removed sections + the
unified line-level diff per modified text file) is going to be reused
by the GitHub-link sync-preview UI on the AdvancedOptionsModal. Pull
it out into a pure component that takes a VersionDiffResponse["diff"]
prop and renders it; keep VersionDiffModal as the version-picker
shell that hands its fetched data to the view.

No behavior change. Same color tokens, same diffLines() chunking,
same per-file binary fallback.
